Revert previous tweak
This reverts commit a2dc2be9acb523036d3f86ff85a9304e9000d6c1.

After looking more closely at the constructor for AdvGame, I realized
that because it really need to be able to load several different files,
it makes more sense to pass in the prefix directly and have it deal with
everything in its constructor. I'll instead update the guide to more
directly address this difference from the Teaching Machine example.